+++ app/models/relation.rb
