+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_07_29_182428) do
+ActiveRecord::Schema.define(version: 2021_07_30_002553) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -58,7 +58,6 @@ ActiveRecord::Schema.define(version: 2021_07_29_182428) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "user_id"
-    t.integer "entity_id"
   end
 
   create_table "relations", force: :cascade do |t|
@@ -76,9 +75,7 @@ ActiveRecord::Schema.define(version: 2021_07_29_182428) do
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
-  add_foreign_key "entities", "groups"
   add_foreign_key "entities", "users"
-  add_foreign_key "groups", "entities"
   add_foreign_key "groups", "users"
   add_foreign_key "relations", "entities"
   add_foreign_key "relations", "groups"
