+++ spec/examples.txt
@@ -1,13 +1,13 @@
 example_id                                | status | run_time        |
 ----------------------------------------- | ------ | --------------- |
-./spec/models/entity_model_spec.rb[1:1:1] | passed | 0.00807 seconds |
-./spec/models/entity_model_spec.rb[1:2:1] | passed | 0.02407 seconds |
-./spec/models/group_model_spec.rb[1:1:1]  | passed | 0.05438 seconds |
-./spec/models/group_model_spec.rb[1:2:1]  | passed | 0.07639 seconds |
-./spec/models/integrations_spec.rb[1:1:1] | passed | 1.39 seconds    |
-./spec/models/integrations_spec.rb[1:1:2] | passed | 6.78 seconds    |
-./spec/models/integrations_spec.rb[1:1:3] | passed | 1.78 seconds    |
-./spec/models/integrations_spec.rb[1:1:4] | passed | 2.11 seconds    |
-./spec/models/integrations_spec.rb[1:1:5] | passed | 2.45 seconds    |
-./spec/models/user_model_spec.rb[1:1:1]   | passed | 0.01376 seconds |
-./spec/models/user_model_spec.rb[1:2:1]   | passed | 0.01194 seconds |
+./spec/models/entity_model_spec.rb[1:1:1] | passed | 0.00709 seconds |
+./spec/models/entity_model_spec.rb[1:2:1] | passed | 0.05142 seconds |
+./spec/models/group_model_spec.rb[1:1:1]  | passed | 0.00787 seconds |
+./spec/models/group_model_spec.rb[1:2:1]  | passed | 0.02315 seconds |
+./spec/models/integrations_spec.rb[1:1:1] | passed | 1.71 seconds    |
+./spec/models/integrations_spec.rb[1:1:2] | passed | 7.26 seconds    |
+./spec/models/integrations_spec.rb[1:1:3] | passed | 1.8 seconds     |
+./spec/models/integrations_spec.rb[1:1:4] | passed | 1.72 seconds    |
+./spec/models/integrations_spec.rb[1:1:5] | passed | 2.37 seconds    |
+./spec/models/user_model_spec.rb[1:1:1]   | passed | 0.02829 seconds |
+./spec/models/user_model_spec.rb[1:2:1]   | passed | 0.05465 seconds |
